requirements.txt
django~=3.2.5
django-imagekit~=4.0.2
pillow~=8.3.1
django-heroku~=0.3.1
gunicorn~=20.1.0
numpy~=1.21.4